AI-backend/requirements.txt
blinker==1.8.2
chardet==5.2.0
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.3.0
MarkupSafe==2.1.5
nltk==3.9.1
pillow==10.4.0
python-docx==1.1.2
regex==2024.9.11
reportlab==4.2.4
textblob==0.18.0.post0
tqdm==4.66.5
typing_extensions==4.12.2
Werkzeug==3.0.4
